Add files via upload
This graph was build in R studio using R programming language and with the aid of ggplot and ggploty.
There is two variation to the graph, one being done with ggplot and the other one being done by ggplotly. In short one is more interactive thanks to ggplotly while the other is more static.
The graph gives us an insight into how the human development is affected when the corruption scale goes higher and which countries are mostly a victim of this.